refactor of scribus and openoffice in document_generic builder- Todo clean import
